// src/app/interfaces/customer.interface.ts
export interface Customer {
    id?: string;
    firstName: string;
    secondName: string;
    firstSurName: string;
    secondSurName: string;
    passport?: string;
    phoneOne: string;
    phoneTwo?: string;
    mainAddressOne: string;
    optionalAddres?: string;
    referencePhoneOne?: string;
    referencePhoneTwo?: string;
    civilStatus: string;
    contactName: string;
    contactPhone: string;
    companyName?: string;
    profession: string;
}